Add files via upload
This commit contains the code for a web app that gives more insight into Party voting, which was the final solution of the BarCamp Hackathon, where I was an assistant programmer and technical architect of the project. This commit also contains the topic analysis of House of Commons data that was a part of a more complex idea that failed to get a high coherence score, leading to a new and final idea.